Many things to talk about.  Local APICs now have their logical destination register (LDR) set along with their Destination Format Register (DFR). This will ensure that flat-model addressing of logical apic values can work.  I don't see why we cannot use physical addressing as well, but for now, it should work just fine.
The IO APIC redirection entries are set up via the values found in the MP table.  This means, they SHOULD be fully functional.  The IO APIC SHOULD
work.  The reason it does not is unknown.  Brian has noticed that Xen's /proc/interrupts does not mention the IO APIC pins.  Therefor Linux is not
using the IO APIC.  His box at home mentions the IO APIC.  Broken motherboard might be an issue!

Also did cool stuff in the vga driver for fun :P  It has testing benefits.  The first two lines will not scroll or be cleared.  Therefore, you can
print a value to the first or second line and it will persist.  Without debugging mode, it will simply retain the XOmB header line.

Notice when you press a key, the keyboard driver status output is different.  The keyboard is enabled before I think it is.  Just for the record.

IRQs 0-15 are mapped somewhere within the range of interrupts 32-48.  IRQ2 does not exist, as it was used as a cascading irq.