Look at that.  SYSCALL\SYSRET ( the mechanism XOmB uses to switch into and out of usermode ) changes cpu flags depending on the R11 register.  We were setting R11 to 0 beforehand, so this was setting the IF flag to 0.  The specification suggests it will be set to 1 anyway, but this is not true.  The specification was updated.  The flags are completely configurable by the OS.  Where is this update?  Not on white paper... in a discussion topic posted by an AMD engineer on the internet: http://www.x86-64.org/pipermail/discuss/2000-October/001019.html.  GREAT.  THANKS AMD.
Interrupts will fire in user mode now.  But, of course, this sucks.  User mode suddenly realizes it cannot access pages belonging to the kernel.  Like the code in kmain.  Looks like we need some
concept of environments and the like so we can give it a page table.  Unless we want to embrace the religion of general protection faults within page fault handlers (is that weird?).

Also, I should note to those working on it that my keyboard driver did in fact work, it is just that interrupts were not enabled in user mode :(  Now all is good...with the exception that nothing
exists in user land.  Yet! haHA!