Some preliminary scheduler stuff.  Some major changes, I suppose.
Francesco DeSensi's context switch code is found in this revision and my commit.  There were weird git errors when he tried to push.  This code is a template meant to be mixin'd where registers need
to be stored on the stack.  It has been implemented within the IDT's isr_common routine, which prettied up that function quite well. (Ignoring the ugly mixin declaration that D has)

Changed the Makefile to chunk together the necessary user runtime code for linking to applications.  We needed to update what constitutes our runtime.  We were linking the entire kernel into the
application!!! This make the kernel, like 3MB instead of around 1.5MB.  Oops.

Added a scaffold for the scheduler within an environment folder for the kernel.  Added environment table stuff and environment descriptors.  I will need to allocate a stack (it just uses the kernel
one currently).  It can use the GRUB modules to execute code within those.

Updated the user application that we use to test to use the syscalls.  It loops an exit syscall, which I believe is the most apt of the syscalls to use.  The kernel now sets up a new environment for
the first (and right now only) environment.  It then, after the kernel starts, SYSRETs into this executable using new syscall functionality.

We should be able to easily support ELF and a new executable type yet to be defined.  ELF is as legacy as the concept of segmented memory.

Now we just need to, oh I dunno, make another application, load it as a grub module, and then implement a scheduler to switch between the two.  The interrupt handler is already in scheduler.d and the
timer is already established (quantum of 50ms).  We are ready to write the damn thing...finally.  We don't need to worry about a runtime or specific and useful syscalls yet.