Some major bug fixes.  It now can context switch from a syscall and an interrupt successfully.  Interrupt stack now uses the register stack for the environment.  TSS now works as expected.  Might be able to leverage the RSP members of the TSS as a better strategy, but it currently uses the IST members.  But the RSP might work better for nested (kernel) interrupts and quickly ascertaining the interrupt origin (user or kernel).
The Code Segment for the kernel is now set as non-conforming.  Otherwise an interrupt would not change the CPL...no matter what the IDT suggests the ring should be to handle the interrupt.  Otherwise, it would fail when using the kernel stack (which is not a usermode accessable page).  I DO NOT KNOW WHY SEGMENTS ARE STILL EXAMINED IN THIS ARCHITECTURE!!!?!?!

Now setting the SS correctly to match what is pushed on the stack by a usermode Interrupt.  Currently, the two environments trigger the scheduler in two different ways.  "application" calls the yield() syscall.  "application2" divides by zero.  (XXX: I have this mapped to the quantum interrupt handler, this is because Bochs does not have an HPET, and I want to test the calling of this procedure.  Have this in mind.  And, please, Don't yell at me. It's a holiday, afterall.  I'm doing my best.)

BUG FIX: GDT.setSysSegment64() , was not setting the base address correctly (only set first 48 bits)  This means the TSS was never valid!  Only figured out when the TSS was finally used, and the kernel crashed immediately.

BUG FIX: TSS address was a physical address, needed to be a virtual address.  Oops.

BUG OUTSTANDING: PIC raises IRQ 0 before being disabled (via masking).  The interrupt remains unacked when the cpu STI occurs.  INT 8 fires.  I am ignoring it, but a comment is there.  This will have to be investigated.  Because INT 8 is the Double Fault interrupt, bad things happen.  I have no clue how to suppress an interrupt other than handling it.

Happy Turkey Day.