Hello,
I have merged my changes, Brian's queue implementation, and Steve's shell.

Auxilary Processors tell the scheduler of their presence and they are marked in a table as a resource.  That is all they do, they are still never used.  We may use this to provide user apps with these processors and take the boot processor to either become another cpu resource, or make it an option to turn on a kernel-only core.  We want to provide both.  It will not be hard to do.

A new mapping is in place.  Four tiers: Kernel shared, CPU specific, Environment-kernel shared, Environment-kernel cpu specific, and then also the user space regions should use up the rest.  Devices and such are still in the same places, but stacks have moved around.

Interrupts should work in the kernel and in userspace.  Weird methods of doing it, but we use sysret when we can, unlike Linux to my knowledge which uses iretq always.  I've tested them pretty well.  We should be aware of the usage of the frame pointer, as any misuse or nonuse of it in the kernel code will break the interrupt mechanism and clobber the stack for the kernel code.

There may be an issue in the schedule() function not scheduling a new environment when only the yield function seems to do so.  The scheduling should be done by the schedule() function which should be called by the yield().  At least that is how I expect it to work.  Brian can update this when he reverts to a priority queue.

The shell polls for a keyboard press.  This can be updated to a keyboard libOS instead of storing the keyboard buffer in the kernel space.  That should limit the keyboard reads to take no context penalty.  Keyboards can be read without switching due to a syscall.  I hope to figure out how this simple kernel-known libOS interface will work soon and have it implemented.  But, um, it works as is for now.

I need to finish the cpu multi-core stuff... should be easy but it takes 25 minutes to boot the kernel via emulation...so yeah...

Should be enough to have concurrent work.

Yay us.