Alright. I added a few things to allow for device pages and buffers to be allocated by the kernel for the environment. The keyboard syscall initKeyboard is implemented. This syscall will allocate two pages to give the user program a keyboard buffer. The first page is the 'readable' page.  This page contains the buffer itself and the write-pointer for the circular buffer. The second is the 'writeable' page, which contains the read-pointer.  The Keyboard libOS, implemented in the libOS directory, implements the 'get' portion of the keyboard buffer and the code to do translations from keyboard scancodes to printable characters.  The grabch() syscall is now completely gone.
All character input from the keyboard is done completely in userland.

Some keyboard scancode sets are required by hardware. My laptop, for instance, defaults to the old set 1. I cannot get the keyboard controller to tell me what scancode set it is currently using. Nor can I use the controller to set the scancode set. (It hangs in VMWare, it would sometimes ignore it in RL anyway).  So, if it outputs random stuff instead of the keys you are pressing...that is currently an issue.  Should be fine in Xen, Bochs, and VMWare which default to the set I am using (set 2) which is the most common code set.

The code sets suggest how keys are mapped to values.
http://www.computer-engineering.org/ps2keyboard/