I couldn't sleep, so I implemented a video driver that runs only in userland. This means no syscall is needed to print to the screen. This also means that the shell won't need to call ANY syscalls. I should get HPET working now, since the program will never voluntarily give up control.
So, now you have a userland Console library to link to. I have pointed Steve's print function to this Console and turned 'echo' (the previous syscall to write to the screen) into 'error', which will print out to the screen regardless of the environment that owns the console. Any kernel writes to the screen will happen in a random place independent of the environment. But, this is ok, a kernel panic would be the only reason for a real-world kernel write to the screen, or a call to 'error', which should only be used for debug and then removed.

It maps the video buffer directly to a read-write page for the environment. The environment handles its own positioning and color information.

This also shows how an environment can link to multiple device resources... the shell uses both this and the keyboard, which are separated.

It needs some locks.

I stepped on Steve's toes again. I rewrote much of your stuff. :/ Apologies. I D-ified your code a little, I suppose. Sorry! But, hey, I also fixed a buffer overflow bug!

I await the malloc and free code. :)

-- Wilks