Added VESA stuff.
VESA uses 16 bit real mode code, so x86 emulation of real mode was necessary. This is added as a package called XOmBemu.

The VESA driver (only working in Bochs) will execute this BIOS code in userland and run the necessary routines to poll the device and gather information about supported modes. It will then pick the 'best' mode and switch to it with the capability of a double buffer.

The driver then writes to the working buffer with simple blitting functions and then flips the buffer. I've added a couple of static arrays that store raw pixel data for our logo for fun.

Signed-off-by: xomboverlord <overlord@xomb.org>