TSS Multicore Fix
When the TSS (Task State Segment) is pointed to by the GDT descriptor, you have to set it as Available. This we already had done. In 64-bit long mode, even though the TSS is never used in terms of it being 'busy', it will set the descriptor to reflect it as 'busy' anyway. When you attempt to set the TSS when it is 'busy' it will GPF. Therefore, if you want to set the TSS via the LTR (Load Task Register) and point it to the GDT's Task Gate Descriptor, you must make sure it is available. To do this, just set the descriptor with every core right before you call LGDT and LTR.

setGDT() {
  SetTaskGate(TASK_GATE_ENTRY, ...)

  asm {
    lgdt (GDT)
    ltr $(TASK_GATE_ENTRY << 3)
  }
}