And so starts the great reorganization of the XOmB source tree.
We're basing things off of the xomb-bare-bones sub-project. Because we're that awesome.

In this commit, we just have the main directory structure, along with the bare bones base stuff. We'll be adding things back in bit by bit, so stay tuned for new updates as we re-work the code base.