Added trampoline code and aux cores now start and enter the kernel.
The trampoline code is copied to the first megabyte of memory (due to
processor cores wanting to start in real mode with its meager address
space). The code itself will take the cpu through from 16 bit real mode
to 32 bit protected and then to the traditional 64 bit bootstrap into
the kernel.

The LocalAPIC sends IPIs to the cores to start them. This is now coded.
The process is odd, but well documented in the Intel manuals.

The cores enter at apEntry in kmain.d.